Compile fixes: move veritomyx.gamma to veritomyx/gamma
